xFunc should return a func w/ identical sig as underlying func
The signature of a func returned by, say, color.PrintFunc should be
the same as fmt.Print. This enables code usage like:

```go
	f := fmt.Print
	if configColor {
		f = color.New(color.FgRed).PrintFunc()
		// f = color.New(color.FgRed).Print // this is sufficient
	}
	f("foo bar")
```

P.S. After this patch, I am not sure why the xFunc family of functions
are still needed (except for backward compatibility). Maybe we should
revise the examples in the documentation to reduce future usage.