Remove `vendor/` folder
I started to like working on projects that **don't** use `vendor/`
folders anymore. The Go ecosystem has improved a lot over the years.
With the introduction of the Go module proxy, it's more efficient and
faster for users as well.

If you're interested, here is a blog post I wrote once that goes into
more details:
https://arslan.io/2019/08/02/why-you-should-use-a-go-module-proxy/